// src/app/models/closer/transactionTextCloser.ts
export class TransactionTextCloser {

   transactionId: string;

   closerCompanyName: string;

   closerCompanyAddress: string;

   closerName: string;

   closerEmail: string;

   closingDate: string;

   closerPhone: string;

   realtorName: string;

   brokerCompanyName: string;

   mls: string;

   propertyAddress: string;

   firstParagraph: string;

   secondParagraph: string;

   thirdParagraph: string;

   fourthParagraph: string;

   fifthParagraph: string;
}